+++ release/apim-server-openapi/build.ts
@@ -124,39 +124,6 @@ const copyAssets = () => {
   console.log(`${logName}: success.`);
 }
 
-const compileSrcs = () => {
-  const funcName = 'compileSrcs';
-  const logName = `${scriptDir}/${scriptName}.${funcName}()`;
-
-  s.cd(`${ReleaseDirBrowser}`);
-  if(s.rm('-rf', `./dist`).code !== 0) process.exit(1);
-  if(s.exec('npx tsc').code !== 0) process.exit(1);
-
-  s.cd(`${ReleaseDirNode}`);
-  if(s.rm('-rf', `./dist`).code !== 0) process.exit(1);
-  if(s.exec('npm install').code !== 0) process.exit(1);
-  if(s.exec('npx tsc').code !== 0) process.exit(1);
-
-  console.log(`${logName}: success.`);
-}
-
-const publishPackages = () => {
-  const funcName = 'publishPackages';
-  const logName = `${scriptDir}/${scriptName}.${funcName}()`;
-
-  const publish = (releaseDir: string) => {
-    s.cd(`${releaseDir}`);
-    if(s.exec('npm publish').code !== 0) process.exit(1);
-    // if(s.exec('npm publish --dry-run').code !== 0) process.exit(1);  
-  }
-
-  publish(ReleaseDirBrowser);
-  publish(ReleaseDirNode);
-
-  console.log(`${logName}: success.`);
-
-}
-
 const main = () => {
   const funcName = 'main';
   const logName = `${scriptDir}/${scriptName}.${funcName}()`;
@@ -166,8 +133,6 @@ const main = () => {
   checkVersions();
   devBuildApimServer();
   copyAssets();
-  compileSrcs();
-  publishPackages()
   console.log(`${logName}: success.`);
 }